pyproject.toml
[tool.poetry]
name = "cxrag-online"
version = "0.1.0"
description = ""
authors = ["ptsadi <putu.adi@hyrd.de>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
google-auth = "^2.37.0"
google-auth-oauthlib = "^1.2.1"
google-auth-httplib2 = "^0.2.0"
pillow = "^11.0.0"
pinecone = "^5.4.2"
numpy = "^2.2.1"
pandas = "^2.2.3"
google-generativeai = "^0.8.3"
streamlit = "^1.41.1"
google-cloud-aiplatform = "^1.75.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
